Fixes #2541. Allows tuning raw MPEG channels in HDHRRecorder.
The HDHomeRun recorder classes were originally based off the HDTVRecorder (at least partially). The HDHRRecorder only tunes ATSC streams. Since most cable operators have many channels which do not contain ATSC info this means that QAM tuning is only likely to work with HDTV channels, assuming the cable operator is obeying the law (a big assumption in some markets). This copies over some of the pertinent bits of code from the DVBRecorder to the HDHRRecorder, and makes hdhrchannel look at the MPEG Program Number if the ATSC channel info is missing.

The change is based on the patch submitted by jwestfall.



git-svn-id: http://svn.mythtv.org/trunk@11521 4b91aca1-2a1f-0410-a73a-a78f2df169e8