Refs #1770. Splits filldata.cpp functions into functional classes.
This looks like a big change, but in fact none of the code had been changed except to move functions and variables to classes (or main).

The Data Direct program insertion function was moved to DataDirectProcessor, the other program data insertion and deletion functions were moved to ProgramData. The xmltv parsing functions were moved to the XMLTVParser class, the channel data update/delete/insert functions were moved to ChannelData, the icon data update/insert functions were moved to IconData, and the overall logic was moved to FillData, while main() was moved to main.cpp and still does the command line parsing and manipulates the classes.

To make the EIT data insertion compatible with XMLTV/DataDirect data I need the program data insertion classes to be accessible to the EIT processor in libmythtv. So the main reason for the splitting is to move the program data insertion methods to libmythtv. At some point in the future it might be useful to have access to the other portions of fill data for the frontend channel editor, but for now I've left the code in mythfilldatabase.

I've tested this with DataDirect and two XMLTV data grabbers. Although since very little code has changed it is more likely that there might be header problems.



git-svn-id: http://svn.mythtv.org/trunk@11619 4b91aca1-2a1f-0410-a73a-a78f2df169e8