Standardizes reporting of ignored channels from the scanner.
Some of the table parsers printed to the console, some to the UI, some not at all. This adds four methods IgnoreDataOnlyMsg(), IgnoreEmptyChanMsg(), IgnoreAudioOnlyMsg() and IgnoreEncryptedMsg() which print the same message when one of these is ignored for PAT, VCT and SDT parsers. The message is printed in English to the console and in the local language (assuming it has been translated) to the UI's textbox.

git-svn-id: http://svn.mythtv.org/trunk@11687 4b91aca1-2a1f-0410-a73a-a78f2df169e8