Fixes #2579. Adds two new system information table types for MPEG streams, "mpeg" and "opencable".
The problem with "OpenCable" is that it uses a different stream id for video 0x80 which is already being used for other data in DVB countries for other uses. We were remapping the 0x80 stream id to the standard video stream id unconditionally which caused problems when this was not in fact a video stream.

With this change we mark a transport as being either a "dvb", "atsc", "opencable" or "mpeg" instead of just marking it as either "dvb" or "atsc" during the channel scan. This allows us to avoid this 0x80->video stream id mapping when in a DVB country, but still perform the remapping when the channel really is an "opencable" channel.

git-svn-id: http://svn.mythtv.org/trunk@11693 4b91aca1-2a1f-0410-a73a-a78f2df169e8