Refs #1866. Refs #2600. Replaces DVBTuning with DTVMultiplex which does not depend on DVB headers and uses the same string parsing routines as the dvb-utils channels.conf reader.
This DTVMultiplex is then used for DTVChannel::Tune() so that the different DTV channels classes no longer require custom tuning code in the channel scanner. This also gets rid of the Linux DVB header dependency in frequencytables.{h,cpp} which simplifies that code as well.

This has been tested with DVB ATSC 8-VSB/QAM-256, DVB-T, DVB-C, ivtv (which uses a DTV capable tuning class), and the HDHomeRun 8-VSB/QAM-256.

git-svn-id: http://svn.mythtv.org/trunk@11702 4b91aca1-2a1f-0410-a73a-a78f2df169e8