Refs #2679. Use SetStrOption() convenience function for setting recording profile options in MPEGRecorder.
This prevents the segfault since the RecorderBase convenience function does a NULL pointer check. I'm still working on the underlying problem which appears to be a problem with Triggered configuration groups that should be saving all their children, so this problem is only happening with new ivtv recorders you add (and probably some other recorders with required options where a null check on byName() is not performed).


git-svn-id: http://svn.mythtv.org/trunk@11725 4b91aca1-2a1f-0410-a73a-a78f2df169e8