Fixes #2622. Uses mythfilldatabase as the default mythfilldatabase program, instead of putting the path in the DB for the default case.
The reason for this is simply to make it easier for people to switch between vendor supplied MythTV in /usr/bin and a locally compiled MythTV in /usr/local/bin, before this change we would keep calling the mythfilldatabase in the originally configured location, even if the installed prefix of the running mythbackend had changed since the setting was originally configured.

Existing path preferences remain unchanged.

git-svn-id: http://svn.mythtv.org/trunk@11729 4b91aca1-2a1f-0410-a73a-a78f2df169e8