When an upcoming recording is on more than one channel with the same callsign, show only one instance of this status on the channel that will be used for recording.
This is another imperfect and possibly confusing approach to this
sticky issue but hopefully better than before. If a show is on a
station that has more than one channel for the same callsign, the
recording font and status WillRecord were shown for each showing
for that callsign in proglist results (upcoming episodes, search
results, search lists, etc.). This is important for Search
Lists->Channels so that a channel list will show that an episode
is going to record even if it is on a different channel number for
the same callsign.  The problem was that this could make it appear
as if the show was scheduled to record twice at the same time for
the two channels.

For example:

1/1 8:00pm 3 KVBC    A Show        1
1/1 8:00pm 123 KVBC  A Show        1

This appears twice but will only record once. Adding the status
column to proglist helped because it shows that this will only
record from an input on card 1. However, I now have cable and
digital cable inputs on both cards 1 and 2 so the card number
does not indicate if the shows will be recorded over s-video from
channel 123 or coax for ch 3.

With this change, only the chanid of the scheduled recording will
be marked with the status of rsWillRecord and other channels are
marked as rsEarilerShowing:

1/1 8:00pm 3 KVBC    A Show        E
1/1 8:00pm 123 KVBC  A Show        1

Now an upcoming episodes list or search result will list each
available showing but will only mark the one channel and card
that the scheduler plans to use.

However, this still leaves the problem of Search Lists->Channels.
Here an execption is made so that the list will always show if an
episode is scheduled to record:

1/1 7:00pm 3 KVBC    Some Show     -
1/1 8:00pm 123 KVBC  A Show        1
1/1 9:00pm 3 KVBC    Another Show  -

Even though this is a list for channel 3 only and "A Show" is
shown on channel 3, the item is overwritten to show that this
KVBC show will be recorded, however, it will be recorded on
card 1 from channel 123.


git-svn-id: http://svn.mythtv.org/trunk@11796 4b91aca1-2a1f-0410-a73a-a78f2df169e8