Fix a bug that was introduced in the Storage Groups patch in the Ringbuffer is_local determination code.  We need to assume a file is local if it starts with a '/', not just if we can find it local using QFile::exists().
Also, as a safety precaution in ProgramInfo::GetPlaybackURL, if the file should
be local (ie, we're on the backend that recorded the file), but we can't find
it, return a filename instead of returning a myth:// URL.  If we return a
myth:// URL, the caller could try to stream a file that doesn't exist on
the backend.

References #2774.



git-svn-id: http://svn.mythtv.org/trunk@12188 4b91aca1-2a1f-0410-a73a-a78f2df169e8