Add a new setting under Setup -> TV SEttings -> Playback to allow the user to force Myth to prefer to stream files from a remote backend.  This preference is ignored by the preview generator when looking for a local file, by mythtranscode since it needs direct access to the file in order to create a transcoded copy, and by the JobQueue when it calls mythtranscode.
The "Hardware Decoder Settings" page has been renamed to "Decoder Settings"
and this new setting has been added to the page.  The "Preferred MPEG2 Decoder"
setting has also been moved to this settings page.

This is a totally different patch, but it was based on the patch by
gdragon at jetcom dt org in #2708.

Closes #2708.



git-svn-id: http://svn.mythtv.org/trunk@12197 4b91aca1-2a1f-0410-a73a-a78f2df169e8