Change how mythburn.py detects whether a file is on a remote filesystem.
The current method of parsing the output of the mount command is a little
fragile and falls over when a mount has a space in the name. It would be
better to use statfs to determine if a file is on a remote fs except that
python doesn't seem to have an equivalent of the 'c' statfs function.

This commit just adds a new function to mytharchivehelper that the script can
call to do the check.

Fixes #2715.

git-svn-id: http://svn.mythtv.org/trunk@12202 4b91aca1-2a1f-0410-a73a-a78f2df169e8