This fixes a problem where exiting the Appearance settings sometimes causes a segfault. It actually only affects the classic menu layout but it was only by luck the DVR and default layouts worked.
Basically this commit makes sure the AppearanceSettings dialog is deleted and
cleaned up before doing the JumpTo("Reload Theme") so that when its detached
from the MythMainWindow it doesn't send an ExitToMainMenuEvent which was 
causing all the problems.

It also only reloads the theme if the user completed the settings dialog by
pressing "Finish".

Fixes #2860, #2580, #2738.

git-svn-id: http://svn.mythtv.org/trunk@12373 4b91aca1-2a1f-0410-a73a-a78f2df169e8