Close #2838 by applying the patch.
Allows the time format sent to the MythShutdownNvramCmd to be specified
rather than using a hardcoded format. Should allow for using commands other
than nvram-wakup to set the wakeup time in the BIOS.

I added a DB update to change the existing MythShutdownNvramCmd settings to 
use the new format so everything should work just the way it was before this 
patch. Also updated the README.


git-svn-id: http://svn.mythtv.org/trunk@12374 4b91aca1-2a1f-0410-a73a-a78f2df169e8