Added ability to request a Preview Image of any size and time within the file using the XML method getPreviewImage.
The new parameters are:

   Width
   Height
   SecsIn

They are all optional.  If only the Width or Height is supplied, the other value will be calculated based on the image aspect ratio.

All images are cached with the name of:

	<image name & ext>.<width>x<height>x<secsIn>.png

The delete logic has been modified to remove all cached images once the video is removed.

Sample request:

	http://mythbackend:6544/getPreviewImage?ChanId=1003&StartTime=2006-12-30T22:30&Width=720&Height=480&SecsIn=120

The test html page in contrib/MythXMLTest has been updated with the new parameters.


git-svn-id: http://svn.mythtv.org/trunk@12396 4b91aca1-2a1f-0410-a73a-a78f2df169e8