Added two new "Duplicate In" options for "Exclude unidentified episodes" and "Exclude old episodes".
The kDupsExGeneric is used in the match phase to exclude matching
showings from a series if the showing does not appear to describe
a specific episode as determined by the 'generic' flag. If set,
only episodes with descriptive information are allowed to record.

"Exclude old episodes" is subtly different than "Record new episodes
only" and the user may prefer either depending on circumstance. Say
there is a series with six showings; two new and four re-runs.
The kDupsNewEpi will match all six and mark four as recstatus "r".
The new kDupsExRepeats will only match the two new showings in the
match phase and no showings will be marked as "r".

The dupin flags were sorted out a bit so that kDupsNewEpi and the
new options have separate bit fields as these are not mutually
exclusive. The current UI is kind of a catch all for one off
specialty options and the interface may change in the future but 
the code is now prepared for setting combinations of dupin flags.


git-svn-id: http://svn.mythtv.org/trunk@12433 4b91aca1-2a1f-0410-a73a-a78f2df169e8