Fixes #2703. Adds PSIP UTC time handling.
 * Adds DVB TDT parsing
 * Adds accurate time estimation based on DVB TDT and ATSC STT time.
 * Passes the accurate time delta to the Conditional Access Module class
 * Adds code to the CAM class to so that it can pass the correct time
   to the CAM wrt to what time the station thinks it is (which can be
   several seconds off from our time even when using NTP.)
 * Moves time handling function to util so it can be shared by the
   new code and the existing time methods in EITHelper.

The DVB spec says that the CAM must be given accurate time WRT to the transmission time of the stream, if either our clock was off or the transmitters time was off then the old code did not do this because we just sent our UTC time to the CAM. With this code we calculate the offset between our time and DVB time and use that to adjust the code we send to the CAM. This change should prevent us from losing more than a second of data when the CAM encryption changes during the transmission of a program we are recording.

The ATSC time offset is just a bonus since the computation is performed in mpegstreamdata, but this and the DVB time offset could be used to set the local clock.



git-svn-id: http://svn.mythtv.org/trunk@12439 4b91aca1-2a1f-0410-a73a-a78f2df169e8