Fixes #2865. Fixes SEGFAULT when -display used without another valid display in DISPLAY env variable.
Unfortunately this requires a MYTH_BINARY_VERSION change so you must recompile & install the plugins after installing the update mythtv libraries.

The problem was that we attempted to open the display before we had parsed and/or set the MythContext's X11 display string. This meant that at best we opened the wrong display when detecting the screen settings the first time, and at worst we segfaulted if the DISPLAY environment variable wasn't set to a valid X11 server.



git-svn-id: http://svn.mythtv.org/trunk@12485 4b91aca1-2a1f-0410-a73a-a78f2df169e8